#6: build: rewrite find_package_local to have extra search paths
The primary problem was that when a <pkg-name>_DIR is set, cmake uses
it in find_package and actually *changes* the value of the variable to
the directory actually found!

Thus,

  find_package_local(kokkos "${kokkos_DIR}/cmake/kokkos" ...)

works properly the first time, but then fails when cmake is
re-run. This bug was very confusing because cmake changed the
behaviour with <pkg-name>_DIR being set.

Thus, find_package_local now takes alterative paths and automaticall
tries multiple until one is found that works:

  find_package_local(
    kokkos "${kokkos_DIR}" Kokkos
    "cmake/Kokkos" "CMake/Kokkos" # New arguments
  )

Additionally, the installed `checkpointConfig.cmake.in` file was
incorrect. It references ${kokkos_DIR} to export the configuration,
causing a failure to load.