#401: alloc: implement/use new in-place alloc wrapper for proper destruction
Implement a new type InPlaceWrapper<T> that wraps an in-place new
allocation and properly destructs it when it goes out of scope. The
expected use is to either call transferOwnership if the allocation is
placed in a managed pointer or letting it go out of scope (this often
happens during copy/move construction implicitly) which will now
properly clean up after itself.